User_2 Flash File
I have a flash file that connect to the same application on the FMS. Each one publishes to a stream called “live”. There are two video objects on the stage that act as a preview of yourself, and the live camera from the other user. (The small video object on the left is your preview, and the larger on the right is the person you are talking to.)Upload the user_2 folder to your server (it doesn’t matter where it go on your server, just be sure you know how to get it using your browser), and browse it.